requirements.txt
pyFluidSynth
Autologging
protobuf